requirements.txt
numpy
pandas
tensorflow
matplotlib
opencv-python
mediapipe
scikit-learn
imageio
tqdm
g2p-en
pronouncing
whisper
ffmpeg-python
g2p
cffi
cycler
decorator
face-alignment
imageio
imageio-ffmpeg
kiwisolver
matplotlib
networkx
numpy
pandas
Pillow
pycparser
pyparsing
python-dateutil
pytz
PyWavelets
PyYAML
scikit-image
scikit-learn
scipy
six
torch
torchvision
tqdm